+++ LibCarla/source/carla/client/TrafficSign.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include "carla/client/Actor.h"
+#include "carla/road/RoadTypes.h"
 
 namespace carla {
 namespace client {
@@ -19,6 +20,9 @@ namespace client {
     const geom::BoundingBox &GetTriggerVolume() const {
       return ActorState::GetBoundingBox();
     }
+
+    carla::road::SignId GetSignId() const;
+
   };
 
 } // namespace client