+++ LibCarla/source/carla/trafficmanager/SimpleWaypoint.h
@@ -78,10 +78,10 @@ namespace traffic_manager {
     uint64_t SetPreviousWaypoint(const std::vector<SimpleWaypointPtr> &next_waypoints);
 
     /// This method is used to set the closest left waypoint for a lane change.
-    void SetLeftWaypoint(SimpleWaypointPtr waypoint);
+    void SetLeftWaypoint(SimpleWaypointPtr &waypoint);
 
     /// This method is used to set the closest right waypoint for a lane change.
-    void SetRightWaypoint(SimpleWaypointPtr waypoint);
+    void SetRightWaypoint(SimpleWaypointPtr &waypoint);
 
     /// This method is used to get the closest left waypoint for a lane change.
     SimpleWaypointPtr GetLeftWaypoint();